Added the links page. Refactored the menu a little bit to give it place.
Added links also to Anuga, Sial, Gulfood and GPC. Fixed the text in a bunch of other links.

Added link to the weather station in the Contacts page.

Added explanation to the maps in the Contacts page and fixed the links so the maps show the proper places for sure.