Added parragraph about Beans and Green Peas in home page.
Corrected messages for inch to mm conversion.

Improved Popcorn page. We now display Size, Explosion and Expansion explanation together with the data about them.
Also, added a photo of the package bag there.

And also photo of packabe bag in Sunflower page.

Improved the layout of the menu so that the picture of the Company page doesn't have the faces of people covered when using the page in phone size.